+++ agent/pkg/server/auth.go
@@ -73,6 +73,7 @@ func formattedServiceAccountPermissionsErrorMessage() string {
 }
 
 type KubeAuthOptions struct {
+	Audiences     []string
 	WorkspaceName types.NamespacedName
 }
 
@@ -139,6 +140,7 @@ func NewKubeAuth(rootLogger *zap.Logger, config *rest.Config, opts KubeAuthOptio
 		authn:         delegatingAuthenticator,
 		authz:         delegatingAuthorizer,
 		workspaceName: opts.WorkspaceName,
+		audiences:     authenticator.Audiences(opts.Audiences),
 	}
 	return a.Authenticate, nil
 }
@@ -148,6 +150,7 @@ type kubeAuth struct {
 	authn         authenticator.Token
 	authz         authorizer.Authorizer
 	workspaceName types.NamespacedName
+	audiences     authenticator.Audiences
 }
 
 // Authenticate implements grpc_auth.AuthFunc to perform authentication and authorization.
@@ -171,7 +174,10 @@ func (a *kubeAuth) Authenticate(ctx context.Context) (context.Context, error) {
 		return nil, err
 	}
 
-	// Authenticate the user.
+	// Authenticate the user via the TokenReview API.
+	// The token is expected to have a specific audience (usually corresponding to a specific workspace),
+	// to prevent a malicious workspace from impersonating the user for any other purpose.
+	ctx = authenticator.WithAudiences(ctx, a.audiences)
 	res, ok, err := a.authn.AuthenticateToken(ctx, token)
 	if err != nil {
 		if apierrors.IsForbidden(err) {